+++ tests/helpers/configureStore.js
@@ -3,7 +3,15 @@ import { createStore, combineReducers } from 'redux';
 export function configureStore() {
     const store = createStore(
             combineReducers({
-                local: localReducer
+                local: localReducer,
+                someGlobalState: (state = { isGlobal: true }, action) => {
+                    switch (action.type) {
+                        case 'SET_GLOBAL':
+                            return Object.assign({}, state, { isGlobal: action.payload });
+                        default:
+                            return state;
+                    }
+                }
             })
         );
     return store;